+++ ts/features/fastLogin/store/selectors/index.ts
@@ -1,11 +1,25 @@
 import { createSelector } from "reselect";
 import { uniqWith, isEqual } from "lodash";
 import { backendStatusSelector } from "../../../../store/reducers/backendStatus";
-import { fastLoginOptIn, fastLoginEnabled } from "../../../../config";
+import {
+  fastLoginOptIn,
+  fastLoginEnabled,
+  isNewCduFlow
+} from "../../../../config";
 import { GlobalState } from "../../../../store/reducers/types";
 import { isPropertyWithMinAppVersionEnabled } from "../../../../store/reducers/featureFlagWithMinAppVersionStatus";
 
-const fastLoginOptInSelector = (state: GlobalState) =>
+export const isEmailUniquenessValidationEnabledSelector = createSelector(
+  backendStatusSelector,
+  backendStatus =>
+    isPropertyWithMinAppVersionEnabled({
+      backendStatus,
+      mainLocalFlag: isNewCduFlow,
+      configPropertyName: "emailUniquenessValidation"
+    })
+);
+
+export const fastLoginOptInSelector = (state: GlobalState) =>
   state.features.loginFeatures.fastLogin.optIn;
 
 /**