+++ ts/sagas/startup.ts
@@ -42,7 +42,7 @@ import {
   watchZendeskSupportSaga
 } from "../features/zendesk/saga";
 import { watchFciSaga } from "../features/fci/saga";
-import { watchWalletV3Saga } from "../features/walletV3/common/saga";
+import { watchWalletSaga as watchWalletV3Saga } from "../features/walletV3/common/saga";
 import I18n from "../i18n";
 import { mixpanelTrack } from "../mixpanel";
 import NavigationService from "../navigation/NavigationService";
@@ -88,7 +88,10 @@ import {
   setProfileHashedFiscalCode
 } from "../store/actions/crossSessions";
 import { handleClearAllAttachments } from "../features/messages/saga/handleClearAttachments";
-import { watchMessageAttachmentsSaga } from "../features/messages/saga";
+import {
+  watchLoadMessageData,
+  watchMessageAttachmentsSaga
+} from "../features/messages/saga";
 import { watchPnSaga } from "../features/pn/store/sagas/watchPnSaga";
 import { startupLoadSuccess } from "../store/actions/startup";
 import { watchIDPaySaga } from "../features/idpay/common/saga";
@@ -335,6 +338,7 @@ export function* initializeApplicationSaga(
     watchMessagePrecondition,
     backendClient.getThirdPartyMessagePrecondition
   );
+  yield* fork(watchLoadMessageData);
   yield* fork(
     watchUpsertMessageStatusAttribues,
     backendClient.upsertMessageStatusAttributes